added scripts to animator

+++ Assets/Scripts/L3-Stage1/MinionBehaviour.cs
@@ -60,7 +60,7 @@ public class MinionBehaviour : Damageable {
         }
     }
 
-    public override void takeDamage(int damage) {
+    public override void takeDamage(float damage) {
         stunTimeRemaining = stunDuration;
     }
 }
